temp/.ipynb_checkpoints/requirements-checkpoint.txt
monai>=1.3.0
nibabel
matplotlib
einops
pytorch-ignite
itk
lmdb
mlflow
pynrrd
clearml
